requirements.txt
# Appium and Selenium dependencies
appium-python-client>=2.11.0
selenium>=4.0.0

# Testing framework
pytest>=7.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=12.0.0

# Logging and reporting
pytest-html>=3.1.0
pytest-metadata>=2.0.0

# Additional utilities
python-dotenv>=0.21.0
requests>=2.28.0

# Code quality (optional)
flake8>=5.0.0
black>=22.0.0
pylint>=2.15.0
